gh-386: Extract API endpoint definition into separate file
